HAB 84 : goal edit
Partial goal edit functionality : A pop up form is displayed with the current name and the user can update the name and set dates. Since it is a few days before deployment Felix and I thought it would be best to play it safe and not try to make new API routes so this method deletes the old goal and makes a new one ( any data in the array or image URLs will not be copied over ). However may be an easy fix if you know how to pass mutiple arguements as props, I will keep trying but not sure if I will get around to it before deployment.